+++ source/tuning_runner/searcher/mcmc_searcher.h
@@ -46,7 +46,7 @@ public:
         originState = currentState = initialState;
         index = initialState;
 
-        for (int i = 0; i < configurations.size(); i++)
+        for (size_t i = 0; i < configurations.size(); i++)
             unexploredIndices.insert(i);
     }
